test(concurrency): dedupe goroutine-bounding harness into shared helper
SonarQube flagged 13.4% duplication on new code (gate <= 3%): the two
bounding tests in lsmkv and roaringsetrange shared an identical ~65-line
sampler/worker/ceiling harness. Extract it into
entities/concurrency/testinghelpers.AssertGoroutineCeiling (same pattern as
adapters/repos/db/vector/testinghelpers) and shrink both tests to setup plus
a do-closure.

The helper intentionally uses bare goroutines (linter exclusion added):
wrapping them in enterrors.GoWrapper would recover panics from the
test-provided closure and let a broken test pass silently.

Both guards keep their properties: 5x green under -race and non-race, and
both still fail when DISABLE_SROAR_MERGE_BUDGET restores the old unbounded
fan-out.
