+++ tools/linter_go_routines.sh
@@ -19,6 +19,7 @@ excluded_files=(
     "usecases/auth/authorization/docs/generator.go" # docs generator
     "tools/dev/generate_release_notes/main.go" # generate release notes tool
     "test/docker/mockoidchelper/mockoidc_helper.go" # used only for OIDC tests
+    "entities/concurrency/testinghelpers/goroutines.go" # test-only helper; bare goroutines keep panics loud in tests
 )
 
 # Check if file is in excluded list
